requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.6
